Upload initial Dialog-flow Project
This is the studybot academy site entry bot. It will register you for the site by collecting your name and an email address so we may send you update and new feature integration news. Once users povide the information the agent will let you into the website. Returning users will be registered and will be allowed into the site after a simple greeting.